requirements.txt
fastapi
openpyxl
pydantic
pywebview
pywin32==303
python-dateutil
uvicorn
